karate-common: QCamera2:HAL1: Do unConfigureCapture before preview restart
Issue:
In non-ZSL, with early preview restart and flash on, back
to back snapshot is not working.

Fix:
In general, unconfigure is called in cancelPicture.
But in early preview restart case, start preview is called
before cancelPicture. Backend is dependent on unconfigure
before starting preview. So, added unconfigure before
starting preview in early restart.

Change-Id: I04116efd7bcd3d5781f6c5b046fead075ae7cb61
CRs-Fixed: 1025509